# Implement specialized MvNormal density based on precision matrix