@ricardoV94 2024-06-25
Final question is just whether we want / can do a similar thing for the MvStudentT. Otherwise it's ready to merge on my end