@ricardoV94 2024-07-12
Last benchmarks, running the following script:

```python
%env OMP_NUM_THREADS=1

USE_TAU = True

import pymc as pm
import numpy as np
import pytensor
import pytensor.tensor as pt

rng = np.random.default_rng(123)

n = 100
L = rng.uniform(low=0.1, high=1.0, size=(n,n))
Sigma = L @ L.T
Q_test = np.linalg.inv(Sigma)
x_test = rng.normal(size=n)

with pm.Model(check_bounds=False) as m:
    Q = pm.Data("Q", Q_test)
    if USE_TAU:
        x = pm.MvNormal("x", mu=pt.zeros(n), tau=Q)
    else:
        x = pm.MvNormal("x", mu=pt.zeros(n), cov=Q)

logp_fn = m.compile_logp().f
logp_fn.trust_input=True
print("logp")
pytensor.dprint(logp_fn)

dlogp_fn = m.compile_dlogp().f
dlogp_fn.trust_input=True
print("dlogp")
pytensor.dprint(dlogp_fn)

%timeit -n 10000 logp_fn(x_test)
%timeit -n 10000 dlogp_fn(x_test)
```

## USE_TAU = TRUE, befare the PR:
```
logp
Composite{((i2 - (i0 * i1)) - i3)} [id A] 'x_logprob' 8
 ├─ 0.5 [id B]
 ├─ DropDims{axis=0} [id C] 7
 │  └─ CAReduce{Composite{(i0 + sqr(i1))}, axis=1} [id D] 6
 │     └─ Transpose{axes=[1, 0]} [id E] 4
 │        └─ SolveTriangular{trans=0, unit_diagonal=False, lower=True, check_finite=True, b_ndim=2} [id F] 2
 │           ├─ Cholesky{lower=True, destructive=False, on_error='nan'} [id G] 1
 │           │  └─ Q [id H]
 │           └─ ExpandDims{axis=1} [id I] 0
 │              └─ x [id J]
 ├─ -91.89385332046727 [id K]
 └─ CAReduce{Composite{(i0 + log(i1))}, axes=None} [id L] 5
    └─ ExtractDiag{offset=0, axis1=0, axis2=1, view=False} [id M] 3
       └─ Cholesky{lower=True, destructive=False, on_error='nan'} [id G] 1
          └─ ···

dlogp
DropDims{axis=1} [id A] '(dx_logprob/dx)' 6
 └─ SolveTriangular{trans=0, unit_diagonal=False, lower=False, check_finite=True, b_ndim=2} [id B] 5
    ├─ Transpose{axes=[1, 0]} [id C] 4
    │  └─ Cholesky{lower=True, destructive=False, on_error='nan'} [id D] 1
    │     └─ Q [id E]
    └─ Neg [id F] 3
       └─ SolveTriangular{trans=0, unit_diagonal=False, lower=True, check_finite=True, b_ndim=2} [id G] 2
          ├─ Cholesky{lower=True, destructive=False, on_error='nan'} [id D] 1
          │  └─ ···
          └─ ExpandDims{axis=1} [id H] 0
             └─ x [id I]
262 µs ± 57.8 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
277 µs ± 54.2 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

## USE_TAU = True after the PR

```
logp
Composite{(i4 * ((i2 - (i0 * i1)) + i3))} [id A] 'x_logprob' 11
 ├─ 2.0 [id B]
 ├─ CAReduce{Composite{(i0 + log(i1))}, axes=None} [id C] 10
 │  └─ ExtractDiag{offset=0, axis1=0, axis2=1, view=False} [id D] 9
 │     └─ Cholesky{lower=True, destructive=False, on_error='raise'} [id E] 8
 │        └─ Q [id F]
 ├─ 183.78770664093454 [id G]
 ├─ DropDims{axis=0} [id H] 7
 │  └─ CGemv{inplace} [id I] 6
 │     ├─ AllocEmpty{dtype='float64'} [id J] 5
 │     │  └─ 1 [id K]
 │     ├─ 1.0 [id L]
 │     ├─ ExpandDims{axis=0} [id M] 4
 │     │  └─ x [id N]
 │     ├─ CGemv{inplace} [id O] 3
 │     │  ├─ AllocEmpty{dtype='float64'} [id P] 2
 │     │  │  └─ Shape_i{1} [id Q] 1
 │     │  │     └─ Q [id F]
 │     │  ├─ 1.0 [id L]
 │     │  ├─ Transpose{axes=[1, 0]} [id R] 'Q.T' 0
 │     │  │  └─ Q [id F]
 │     │  ├─ x [id N]
 │     │  └─ 0.0 [id S]
 │     └─ 0.0 [id S]
 └─ -0.5 [id T]

dlogp
CGemv{inplace} [id A] '(dx_logprob/dx)' 5
 ├─ CGemv{inplace} [id B] 4
 │  ├─ AllocEmpty{dtype='float64'} [id C] 3
 │  │  └─ Shape_i{0} [id D] 2
 │  │     └─ Q [id E]
 │  ├─ 1.0 [id F]
 │  ├─ Q [id E]
 │  ├─ Mul [id G] 1
 │  │  ├─ [-0.5] [id H]
 │  │  └─ x [id I]
 │  └─ 0.0 [id J]
 ├─ -0.5 [id K]
 ├─ Transpose{axes=[1, 0]} [id L] 'Q.T' 0
 │  └─ Q [id E]
 ├─ x [id I]
 └─ 1.0 [id F]
160 µs ± 34.1 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
15.9 µs ± 2.29 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

## For reference: USE_TAU = False before and after (unchanged)
Before:
```
...
260 µs ± 13 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
275 µs ± 19.5 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

After:
```
...
259 µs ± 46.3 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
275 µs ± 30.6 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```