@ricardoV94 2024-07-12
`numpyro` tests are failing probably because it now requires the more recent versions of jax. should be fixed by #7407