@ricardoV94 2024-08-03
> Looks good. I think it is possible that we could further improve the MvNormal in both parametrizations, but this is definetly an improvement as it is.
> Most of all I think we should do the same for the CholeskyMvNormal. Looks like we are just computing the cov just to re-compute the cholesky again? At some point we did make use of the cholesky directly, but I guess that got lost in a refactor wtih the pytensor RVs?

We don't recompute the Cholesky, we have rewrites to remove it and even a specific test for it: https://github.com/pymc-devs/pymc/blob/b407c01accc65a68d17803b6208d6dfdb0e40877/tests/distributions/test_multivariate.py#L2368